Added a cs.js file, it doesn't work btw.
I was going to try to set up js so that the header of each page was the same (since we have it typed out on each html page, it would save time whenever we alter it).  I ended up not being able to.
Then I wanted to make it so that the cs.html file would output information from the json file, and i also coudlnt. Made a cs,js file so that it would be more organized, but the code on it still doesn't work.